refactor: move s2i build service factory and implementation into their own package.
