+++ annotations/s2i-annotations/src/main/resources/META-INF/services/io.dekorate.BuildServiceFactory
@@ -1,1 +1,1 @@
-io.dekorate.s2i.S2iBuildServiceFactory
+io.dekorate.s2i.buildservice.S2iBuildServiceFactory
